# packages/shared-types/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: "http://localhost:3000/api/v1" | (string & {});
};

export type AccessToken = {
  /**
   * JWT access token
   */
  accessToken: string;
};

export type User = {
  id?: string;
  email?: string;
  createdAt?: string;
};

export type AuthResponse = User & AccessToken;

export type Task = {
  id?: string;
  title?: string;
  description?: string | null;
  status?: "TODO" | "IN_PROGRESS" | "DONE";
  priority?: "LOW" | "MEDIUM" | "HIGH";
  dueDate?: string | null;
  createdAt?: string;
  updatedAt?: string;
};

export type PaginatedTasks = {
  items: Array<Task>;
  page: number;
  limit: number;
  total: number;
};

export type ErrorResponse = {
  message: string;
};

export type PostAuthRegisterData = {
  body: {
    email: string;
    password: string;
  };
  path?: never;
  query?: never;
  url: "/auth/register";
};

export type PostAuthRegisterErrors = {
  /**
   * Validation error
   */
  400: ErrorResponse;
  /**
   * User already exists
   */
  409: unknown;
};

export type PostAuthRegisterError =
  PostAuthRegisterErrors[keyof PostAuthRegisterErrors];

export type PostAuthRegisterResponses = {
  /**
   * User registered
   */
  201: AuthResponse;
};

export type PostAuthRegisterResponse =
  PostAuthRegisterResponses[keyof PostAuthRegisterResponses];

export type PostAuthLoginData = {
  body: {
    email: string;
    password: string;
  };
  path?: never;
  query?: never;
  url: "/auth/login";
};

export type PostAuthLoginErrors = {
  /**
   * Invalid credentials
   */
  400: unknown;
  /**
   * User not found
   */
  404: unknown;
};

export type PostAuthLoginResponses = {
  /**
   * Login successful
   */
  200: AuthResponse;
};

export type PostAuthLoginResponse =
  PostAuthLoginResponses[keyof PostAuthLoginResponses];

export type PostAuthRefreshData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/auth/refresh";
};

export type PostAuthRefreshErrors = {
  /**
   * Invalid or expired refresh token
   */
  401: unknown;
};

export type PostAuthRefreshResponses = {
  /**
   * Access token refreshed
   */
  200: AccessToken;
};

export type PostAuthRefreshResponse =
  PostAuthRefreshResponses[keyof PostAuthRefreshResponses];

export type GetTasksData = {
  body?: never;
  path?: never;
  query?: {
    status?: "TODO" | "IN_PROGRESS" | "DONE";
    priority?: "LOW" | "MEDIUM" | "HIGH";
    search?: string;
    sortBy?: "createdAt" | "dueDate";
    order?: "asc" | "desc";
    page?: number;
    limit?: number;
  };
  url: "/tasks";
};

export type GetTasksResponses = {
  /**
   * Paginated tasks list
   */
  200: PaginatedTasks;
};

export type GetTasksResponse = GetTasksResponses[keyof GetTasksResponses];

export type PostTasksData = {
  body: {
    title: string;
    description?: string;
    priority?: "LOW" | "MEDIUM" | "HIGH";
    status?: "TODO" | "IN_PROGRESS" | "DONE";
    dueDate?: string;
  };
  path?: never;
  query?: never;
  url: "/tasks";
};

export type PostTasksResponses = {
  /**
   * Task created
   */
  201: Task;
};

export type PostTasksResponse = PostTasksResponses[keyof PostTasksResponses];

export type DeleteTasksByIdData = {
  body?: never;
  path?: {
    id?: string;
  };
  query?: never;
  url: "/tasks/{id}";
};

export type DeleteTasksByIdResponses = {
  /**
   * Task deleted
   */
  204: void;
};

export type DeleteTasksByIdResponse =
  DeleteTasksByIdResponses[keyof DeleteTasksByIdResponses];

export type GetTasksByIdData = {
  body?: never;
  path: {
    id: string;
  };
  query?: never;
  url: "/tasks/{id}";
};

export type GetTasksByIdResponses = {
  /**
   * Task details
   */
  200: Task;
};

export type GetTasksByIdResponse =
  GetTasksByIdResponses[keyof GetTasksByIdResponses];

export type PutTasksByIdData = {
  body: {
    title?: string;
    description?: string;
    status?: "TODO" | "IN_PROGRESS" | "DONE";
    priority?: "LOW" | "MEDIUM" | "HIGH";
    dueDate?: string;
  };
  path?: never;
  query?: never;
  url: "/tasks/{id}";
};

export type PutTasksByIdResponses = {
  /**
   * Task updated
   */
  200: Task;
};

export type PutTasksByIdResponse =
  PutTasksByIdResponses[keyof PutTasksByIdResponses];